% This is a template R document file
% Distributed under GPL 3 or later

\name{plot.CDF}
\alias{plot.CDF}
\title{Plot a CDF Object}
\description{
  To plot a ``CDF'' object generated by \code{kCDF(...)}.
}
\usage{
 \method{plot}{CDF}(x, CI=TRUE, alpha=0.05, main = NULL, xlab = NULL, ylab = "CDF", lwd=2, lty=1, ...)
}
\arguments{
\item{x}{a ``CDF'' object generated by \code{kCDF(...)}.}
\item{CI}{If TRUE, the pointwise confidence intervals will be plotted.}
\item{alpha}{the significant level. The default is 0.05 which generates 95\% confidence intervals for the CDF.}
\item{main}{see \code{\link{par}}}
\item{xlab}{see \code{\link{par}}}
\item{ylab}{see \code{\link{par}}}
\item{lwd}{see \code{\link{par}}}
\item{lty}{see \code{\link{par}}}
\item{...}{further arguments for the plot function.}
}
\details{
This function is to plot the estimated function generated by \code{kCDF(...)}}
\author{
  X.F. Wang \email{wangx6@ccf.org}
}
\seealso{
 \code{\link{kCDF}}.
}
\keyword{smoothing}