% This is a template R document file
% Distributed under GPL 3 or later

\name{plot.ROC}
\alias{plot.ROC}
\title{Plot a ROC Object}
\description{
  To plot a ``ROC'' object generated by \code{kROC(...)}.
}
\usage{
 \method{plot}{ROC}(x, main = NULL, diagonal = TRUE, xlab = "FPR", ylab = "TPR", type = "l", lwd=2, ...)
}
\arguments{
\item{x}{a ``ROC'' object generated by \code{kROC(...)}.}
\item{diagonal}{if TRUE, the diagonal line will be plotted.}
\item{main}{see \code{\link{par}}}
\item{xlab}{see \code{\link{par}}}
\item{ylab}{see \code{\link{par}}}
\item{type}{see \code{\link{par}}}
\item{lwd}{see \code{\link{par}}}
\item{...}{further arguments for the plot function.}
}
\details{
This function is to plot the estimated function generated by \code{kROC(...)}}
\author{
  X.F. Wang \email{wangx6@ccf.org}
}
\seealso{
 \code{\link{kROC}}.
}
\keyword{smoothing}