% This is a template R document file
% Distributed under GPL 3 or later

\name{print.AUC}
\alias{print.AUC}
\title{Print a AUC Object}
\description{
  To print a ``AUC'' object generated by \code{AUC(...)}.
}
\usage{
 \method{print}{AUC}(x, digits = NULL, ...)
}
\arguments{
\item{x}{a ``AUC'' object generated by \code{AUC(...)}.}
\item{digits}{integer indicating the number of decimal places to be used.}
\item{...}{further arguments for the print function.}
}
\details{
This function is to print the summary description from the object generated by \code{AUC(...)}}
\author{
  X.F. Wang \email{wangx6@ccf.org}
}
\seealso{
 \code{\link{AUC}}.
}
\keyword{smoothing}

