% This is a template R document file
% Distributed under GPL 3 or later

\name{print.CDF}
\alias{print.CDF}
\title{Print a CDF Object}
\description{
  To print a ``CDF'' object generated by \code{kCDF(...)}.
}
\usage{
 \method{print}{CDF}(x, digits = NULL, ...)
}
\arguments{
\item{x}{a ``CDF'' object generated by \code{kCDF(...)}.}
\item{digits}{integer indicating the number of decimal places to be used.}
\item{...}{further arguments for the print function.}
}
\details{
This function is to print the summary description from the object generated by \code{kCDF(...)}}
\author{
  X.F. Wang \email{wangx6@ccf.org}
}
\seealso{
 \code{\link{kCDF}}.
}
\keyword{smoothing}

