% This is a template R document file
% Distributed under GPL 3 or later

\name{print.ROC}
\alias{print.ROC}
\title{Print a ROC Object}
\description{
  To print a ``ROC'' object generated by \code{kROC(...)}.
}
\usage{
 \method{print}{ROC}(x, digits = NULL, ...)
}
\arguments{
\item{x}{a ``ROC'' object generated by \code{kROC(...)}.}
\item{digits}{integer indicating the number of decimal places to be used.}
\item{...}{further arguments for the print function.}
}
\details{
This function is to print the summary description from the object generated by \code{kROC(...)}}
\author{
  X.F. Wang \email{wangx6@ccf.org}
}
\seealso{
 \code{\link{kROC}}.
}
\keyword{smoothing}

